WIP: getting view_builder wired in
We want to use `MviewMetadata` as a default class for the health check.
That requires a `RemoteRecord` model, which needs to be backed by an
Oracl view. That means we need the `view_builder` gem and its
configuration. Currently getting:

```ruby
Failure/Error: ActiveRecord::Base.establish_connection(DatabaseTasks.database_configuration[DatabaseTasks.env])

NameError:
  uninitialized constant ActiveRecord::Tasks::DatabaseTasks::Rails
```